moves Exhibits::Bibliography -> Bibliography
